pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"


# Can be replaced to [project] when poetry-dynamic-versioning drops [tool.poetry] support in favor of [project] table.
[tool.poetry]
name = "platform-buckets-api"
version = "0.0.0"
description = "Apolo platform buckets api"
readme = "README.md"
license = "Proprietary"
authors = ["Apolo.us"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
aiohttp = "3.13.0"
neuro-auth-client = "25.8.2"
neuro-logging = "25.10.0"
marshmallow = "3.26.1"
apispec = "6.8.4"
aiohttp-apispec = "3.0.0b2"
markupsafe = "3.0.3"
aiobotocore = "2.25.0"
bmc = "0.0.4"
azure-storage-blob = "12.26.0"
google-cloud-storage = "3.4.1"
google-cloud-iam = "2.19.1"
google-api-python-client = "2.184.0"
apolo-kube-client ="25.10.7"
apolo-events-client = "25.8.0"

[tool.poetry.scripts]
platform-buckets-api = "platform_buckets_api.api:main"

[tool.poetry.group.dev.dependencies]
mypy = "1.18.2"
pre-commit = "4.3.0"
codecov = "2.1.13"
docker = "7.1.0"
pytest = "8.4.2"
pytest-asyncio = "1.2.0"
pytest-cov = "7.0.0"
apolo-events-client = {extras = ["pytest"], version = "25.8.0"}

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"
poetry-dynamic-versioning = { version = ">=1.0.0,<2.0.0", extras = ["plugin"] }

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"

[tool.isort]
profile = "black"
combine_as_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning:jose",
    "ignore::ResourceWarning",
]

[tool.coverage.run]
branch = true
source = ["platform_buckets_api"]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_untyped_defs = true
follow_imports = "silent"
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "aiohttp_security"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "aiohttp_security.api"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "jose"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "neuro_auth_client.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pytest"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "setuptools"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "trafaret"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "aiohttp_apispec"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "docker.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "aiobotocore.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "botocore.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "bmc.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "azure.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "google.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "googleapiclient.*"
ignore_missing_imports = true

[tool.black]
exclude = '''
/(
      \.git
    | venv
    | __pycache__
)/
'''
